Add CustomerRepository with initialize code, remove DatabaseInitializer
