+++ chapter5/src/main/kotlin/com/microservices/chapter5/CustomerRepository.kt
@@ -1,5 +1,21 @@
 package com.microservices.chapter5
 
-import org.springframework.data.repository.reactive.ReactiveCrudRepository
+import jakarta.annotation.PostConstruct
+import org.springframework.data.mongodb.core.ReactiveMongoTemplate
+import org.springframework.stereotype.Repository
+import reactor.core.publisher.Mono
+import reactor.kotlin.core.publisher.toMono
 
-interface CustomerRepository : ReactiveCrudRepository<Customer, Int>
+@Repository
+class CustomerRepository(private val template: ReactiveMongoTemplate) {
+    companion object {
+        val initialCustomers = listOf(
+            Customer(1, "Kotlin"),
+            Customer(2, "Spring"),
+            Customer(3, "Microservice", Customer.Telephone("+44", "7123456789"))
+        )
+    }
+    @PostConstruct
+    fun initializeRepository() = initialCustomers.map(Customer::toMono).map(this::create).map(Mono<Customer>::subscribe)
+    fun create(customer: Mono<Customer>) = template.save(customer)
+}
